+++ tests/integration/plugins/test_npm.py
@@ -185,9 +185,8 @@ def test_npm_self_contained(new_dir, partitions):
         f"""\
         parts:
           hello-dep:
-            plugin: npm
+            plugin: npm-use
             source: {new_dir / "hello-dep-v1"}
-            npm-publish-to-cache: true
             build-attributes:
               - self-contained
           hello-app:
@@ -239,29 +238,25 @@ def test_npm_self_contained_version_resolution(new_dir, partitions):
         f"""\
         parts:
           hello-dep-v1:
-            plugin: npm
+            plugin: npm-use
             source: {new_dir / "hello-dep-v1"}
-            npm-publish-to-cache: true
             build-attributes:
               - self-contained
           hello-dep-v1-1:
-            plugin: npm
+            plugin: npm-use
             source: {new_dir / "hello-dep-v1.1"}
-            npm-publish-to-cache: true
             after:
               - another-dep-v2
             build-attributes:
               - self-contained
           hello-dep-v2:
-            plugin: npm
+            plugin: npm-use
             source: {new_dir / "hello-dep-v2"}
-            npm-publish-to-cache: true
             build-attributes:
               - self-contained
           another-dep-v2:
-            plugin: npm
+            plugin: npm-use
             source: {new_dir / "another-dep-v2"}
-            npm-publish-to-cache: true
             build-attributes:
               - self-contained
           hello-app: