+++ craft_parts/parts.py
@@ -138,7 +138,8 @@ class PartSpec(BaseModel):
     During the pull step, the part fetches the repository from the specified commit
     up to the target commit, the target tag, or the tip of the target branch.
 
-    In conjunction with ``source-commit``, setting this key allows for a shallow clone.
+    If this value is non-zero and ``source-commit`` is set to a full-length commit hash,
+    a shallow clone is pulled.
 
     Equivalent to the
     :literalref:`--depth<https://git-scm.com/docs/git-pull#Documentation/git-pull.txt---depthdepth>`